requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
whitenoise==6.7.0
